requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# 얼굴 감정 분석
fer==22.5.1
tensorflow==2.20.0

# 얼굴 랜드마크 (눈썹 찡그림 / EAR / 고개 기울기)
mediapipe==0.10.33

# 이미지 처리 (Python 3.13 호환 버전으로 고정 해제)
opencv-python-headless
numpy>=2.0.0

# GPT 판정
openai>=1.30.0

# 환경변수 로드
python-dotenv>=1.0.0
